[UPDATE] BJS2_68686 Убрал тесты из общего мейна и перенёс во внутренний мейн, чтобы избежать конфликта.
